+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2018-08-07
 
+- require asset classes for taskdef registration
 - register executor, asset, taskdef
 
 2018-08-06